Add deferred init handler list
This solves a problem evolved from `u.ie9.js`: `document.body` was not defined
when the script was executed. But it is available after `DOMContentLoaded`.
However, the user may add init handler with u(func) at any time. But the
specialized class function have to be overridden *before* any handler is called.
The ordered initializer list solves this problem.